requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
asyncssh==2.15.0
atpublic==5.0
attrs==24.1.0
billiard==4.2.0
boto3==1.34.156
botocore==1.34.156
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
cryptography==43.0.0
debugpy==1.8.3
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.1
dvc==3.53.0
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.4.0
entrypoints==0.4
executing==2.0.1
filelock==3.15.4
flatten-dict==0.4.2
flufl.lock==8.1.0
frozenlist==1.4.1
fsspec==2024.6.1
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
google-auth-oauthlib==1.2.1
googlemaps==4.10.0
grandalf==0.8
greenlet==3.0.3
gspread==6.1.2
gto==1.7.1
httplib2==0.22.0
hydra-core==1.3.2
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
iterative-telemetry==0.0.8
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.3.7
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.0.1
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
platformdirs==3.11.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
pygit2==1.15.1
Pygments==2.18.0
pygtrie==2.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
scmrepo==3.3.6
semver==3.0.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sql==2022.4.0
SQLAlchemy==2.0.31
sqltrie==0.11.1
stack-data==0.6.3
tabulate==0.9.0
tomlkit==0.13.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
yarl==1.9.4
zc.lockfile==3.0.post1
